# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/api/auth/signin": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
...
    "/api/companies/{id}": {
      "patch": {
        "operationId": "CompaniesController_updateCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Companies"]
      },
      "delete": {
        "operationId": "CompaniesController_deleteCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Companies"]
      },
      "get": {
        "operationId": "CompaniesController_getCompanyById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Companies"]
      }
    },
...
    "/api/carts/active": {
      "get": {
        "operationId": "CartsController_getUserActiveCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User active cart returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the active cart for current user",
        "tags": ["Carts"]
      }
    },
    "/api/carts/populated": {
      "get": {
        "operationId": "CartsController_getPopulatedCarts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Populated carts list returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cart"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get carts with populated related data for current user",
        "tags": ["Carts"]
      }
    },
    "/api/carts/summary": {
      "get": {
        "operationId": "CartsController_getSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart summary returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get summary of user carts",
        "tags": ["Carts"]
      }
    },
    "/api/carts": {
      "post": {
        "operationId": "CartsController_createCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new cart",
        "tags": ["Carts"]
      },
      "patch": {
        "operationId": "CartsController_updateCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the user's cart partially",
        "tags": ["Carts"]
      }
    },
    "/api/carts/items": {
      "post": {
        "operationId": "CartsController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to user's cart",
        "tags": ["Carts"]
      }
    },
    "/api/carts/items/{productId}": {
      "delete": {
        "operationId": "CartsController_removeItem",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "ID of the product to remove",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove item from user's cart",
        "tags": ["Carts"]
      }
    },
    "/api/carts/clear": {
      "delete": {
        "operationId": "CartsController_clearCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear all items from user's cart",
        "tags": ["Carts"]
      }
    },
    "/api/carts/checkout": {
      "post": {
        "operationId": "CartsController_checkout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart checked out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Checkout the user's cart",
        "tags": ["Carts"]
      }
    },
...
    "/api/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Categories"]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Categories"]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Categories"]
      }
    },
    "/api/categories/{id}/status": {
      "patch": {
        "operationId": "CategoriesController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Categories"]
      }
    },
    "/api/categories/parent/{parentId}": {
      "get": {
        "operationId": "CategoriesController_findByParentId",
        "parameters": [
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Categories"]
      }
    }
  },
  "info": {
    "title": "practice",
    "description": "API documentation for my practice app",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Iranian phone number of the user",
            "example": "0989123456789"
          },
          "nationalId": {
            "type": "string",
            "description": "Iranian national ID (meli code)",
            "example": "2284280072"
          }
        },
        "required": ["phoneNumber", "nationalId"]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Iranian phone number of the user",
            "example": "+989123456789"
          }
        },
        "required": ["phoneNumber"]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Iranian phone number of the user",
            "example": "+989123456789"
          },
          "otp": {
            "type": "string",
            "description": "One-time password (OTP) for verification",
            "example": "1234"
          }
        },
        "required": ["phoneNumber", "otp"]
      },
      "Cart": {
        "type": "object",
        "properties": {}
      },
...
      "CartItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "MongoDB ObjectId of the product",
            "example": "507f1f77bcf86cd799439011"
          },
          "companyId": {
            "type": "string",
            "description": "MongoDB ObjectId of the supplier company",
            "example": "507f1f77bcf86cd799439011"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the product in the cart",
            "example": 2,
            "minimum": 1
          },
          "priceAtAdd": {
            "type": "number",
            "description": "Price of the product at the time of adding to cart (in IRR)",
            "example": 2000000,
            "minimum": 0
          },
          "variant": {
            "type": "object",
            "description": "Selected variant of the product (e.g., size, packaging)",
            "example": {
              "name": "Packaging",
              "value": "50 kg"
            }
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for the cart item",
            "example": "Urgent delivery required"
          },
          "discount": {
            "type": "object",
            "description": "Discount applied to the cart item",
            "example": {
              "type": "percentage",
              "value": 10
            }
          }
        },
        "required": ["productId", "companyId", "quantity", "priceAtAdd"]
      },
      "CreateCartDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "MongoDB ObjectId of the user owning the cart",
            "example": "507f1f77bcf86cd799439011"
          },
          "items": {
            "description": "List of items in the cart",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "description": "Total amount of the cart (in IRR)",
            "example": 4000000,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "Status of the cart",
            "enum": ["active", "abandoned", "checked_out"],
            "example": "active"
          }
        },
        "required": ["userId", "items"]
      },
      "PartialType": {
        "type": "object",
        "properties": {}
      }
    }
  }
}